+++ lib/logstorage/filter_test.go
@@ -207,7 +207,7 @@ func testFilterMatchForStorage(t *testing.T, s *Storage, tenantID TenantID, f fi
 	var results []result
 
 	const workersCount = 3
-	s.search(workersCount, so, qs, nil, func(_ uint, br *blockResult) {
+	s.searchParallel(workersCount, so, qs, nil, func(_ uint, br *blockResult) {
 		// Verify columns
 		cs := br.getColumns()
 		if len(cs) != 2 {